DND messaging fixes, update atsume

+++ dnd_messaging/commands.py
@@ -10,22 +10,6 @@ from .ai import score_message, rewrite_failed_message, message_to_log
 from .models import *
 
 
-
-@tanjun.as_message_command("dnd_message", "hello", "hey", "howdy")
-async def dnd_message(
-        ctx: tanjun.abc.Context,
-) -> None:
-    stats = await DNDStats.objects.get_or_create(user=ctx.member.id, _defaults={
-        "charisma": roll_stat(),
-        "intelligence": roll_stat(),
-        "wisdom": roll_stat(),
-        "deception": roll_stat(),
-        "humor": roll_stat()
-    })
-    print(stats)
-    await ctx.respond(f"hello {stats}")
-
-
 @tanjun.as_message_command("dnd_stats")
 async def set_dnd_stats(ctx: tanjun.abc.Context):
     channel, created = await DNDStatsChannel.objects.get_or_create(guild=ctx.guild_id,
@@ -83,12 +67,15 @@ async def on_message(event: hikari.events.MessageCreateEvent, client: alluka.Inj
     if event.author.is_bot:
         return
 
+    stats = await generate_member_stats(client, event.message.guild_id, event.message.member)
+    if stats is None:
+        return
+
     member = await client.rest.fetch_member(event.message.guild_id, event.message.author)
     channel = cast(hikari.GuildTextChannel, await event.message.fetch_channel())
     await event.message.delete()
 
     async with channel.trigger_typing():
-        stats = await generate_member_stats(client, event.message.guild_id, event.message.member)
         print(stats)
 
         log = await message_to_log(client, channel)
@@ -107,12 +94,12 @@ async def on_message(event: hikari.events.MessageCreateEvent, client: alluka.Inj
         if user_stat + dice_roll < score or dice_roll == 1:
             new_message = await rewrite_failed_message(log, member, message, category)
 
-            new_message = f"-# {abbr_cat} - rolled a {dice_roll} - failure!\n{new_message}"
+            new_message = f"-# {abbr_cat} {score} - rolled a {dice_roll} - failure!\n{new_message}"
 
             await fake_message(client, channel, member, new_message)
             return
         else:
-            new_message = f"-# {abbr_cat} - rolled a {dice_roll} - success!\n{message}"
+            new_message = f"-# {abbr_cat} {score} - rolled a {dice_roll} - success!\n{message}"
 
         await fake_message(client, channel, member, new_message)
